requirements.txt
openai==2.5.0
python-dotenv==1.1.1
